requirements.txt
Django>=4.2
djangorestframework
numpy
pillow
tensorflow
opencv-python
scikit-learn